Refactor query replay -
1) implement using Observables
2) add missing type defs
