+++ public/files/js/models/query/switchmc.ts
@@ -19,11 +19,13 @@
  */
 
 import * as Immutable from 'immutable';
-import RSVP from 'rsvp';
 import {AjaxResponse} from '../../types/ajaxResponses';
 import {StatefulModel} from '../base';
 import {PageModel} from '../../app/page';
 import { Action, IFullActionControl } from 'kombo';
+import { Observable } from 'rxjs';
+import { tap, map } from 'rxjs/operators';
+import { Kontext } from '../../types/common';
 
 
 export interface SwitchMainCorpFormProperties {
@@ -69,20 +71,28 @@ export class SwitchMainCorpModel extends StatefulModel {
         return this.layoutModel.createActionUrl('switch_main_corp', args);
     }
 
-    syncFrom(fn:()=>RSVP.Promise<AjaxResponse.SwitchMainCorpArgs>):RSVP.Promise<AjaxResponse.SwitchMainCorpArgs> {
-        return fn().then(
-            (data) => {
-                if (data.form_type === 'switchmc') {
-                    this.maincorpValues = this.maincorpValues.set(data.op_key, data.maincorp);
-                    return data;
-
-                } else if (data.form_type === 'locked') {
-                    return null;
-
-                } else {
-                    throw new Error('Cannot sync switchmc model - invalid form data type: ' + data.form_type);
+    syncFrom(src:Observable<AjaxResponse.SwitchMainCorpArgs>):Observable<AjaxResponse.SwitchMainCorpArgs> {
+        return src.pipe(
+            tap(
+                (data) => {
+                    if (data.form_type === Kontext.ConcFormTypes.SWITCHMC) {
+                        this.maincorpValues = this.maincorpValues.set(data.op_key, data.maincorp);
+                    }
                 }
-            }
+            ),
+            map(
+                (data) => {
+                    if (data.form_type === Kontext.ConcFormTypes.SWITCHMC) {
+                        return data;
+
+                    } else if (data.form_type === 'locked') {
+                        return null;
+
+                    } else {
+                        throw new Error('Cannot sync switchmc model - invalid form data type: ' + data.form_type);
+                    }
+                }
+            )
         );
     }
