+++ public/files/js/pages/freq.ts
@@ -28,7 +28,6 @@ import {MLFreqFormModel, TTFreqFormModel, FreqFormInputs, FreqFormProps} from '.
 import {Freq2DTableModel} from '../models/freqs/ctable';
 import {Freq2DFlatViewModel} from '../models/freqs/flatCtable';
 import {CTFormProperties, CTFormInputs, Freq2DFormModel} from '../models/freqs/ctFreqForm';
-import {IndirectQueryReplayModel} from '../models/query/replay';
 import {QuerySaveAsFormModel} from '../models/query/save';
 import {fetchQueryFormArgs} from '../models/query/first';
 import {init as freqFormFactory} from '../views/freqs/forms';
@@ -43,6 +42,7 @@ import {ConfIntervals, DataPoint} from '../charts/confIntervals';
 import {TextTypesModel} from '../models/textTypes/main';
 import {NonQueryCorpusSelectionModel} from '../models/corpsel';
 import { KontextPage } from '../app/main';
+import { IndirectQueryReplayModel } from '../models/query/replay/indirect';
 
 declare var require:any;
 // weback - ensure a style (even empty one) is created for the page
